🎨 introduce file to handle the devices
